Finished Second Scenario
Added extensibility to the game by using a queue in case there is a need to add more that one question after each scenario. Also finished the second scenario and changed the mainUI script to check if there are any questions before showing the questionnaire or just let the player continue on the next scenario.