Completed up to sixth game
Updated some scripts for better extendability mostly the script that handles the mixing bowl which is now split on 2 scripts one handling the animation and one the functionality of it. I also update some of the prefabs so they hold more data that is reused like the colliders and scripts assigned to them.